Indexer: extract indexing each content type into its own class.
Indexer still isn't perfect, but hey, what a nice and welcome change.
